chore: remove SwapModalFooterMod and move logic to index
